Restructures IPC mempool tests into a subdirectory.
